feat: send position of all players via query

+++ lib/render.c
@@ -5,7 +5,7 @@ void RenderRectangle(ecs_iter_t *it) {
   Position *positions = ecs_field(it, Position, 0);
   RenderableRectangle *rectangles = ecs_field(it, RenderableRectangle, 1);
 
-  for (size_t i = 0; i < it->count; i++) {
+  for (int i = 0; i < it->count; i++) {
     Position *p = &positions[i];
     RenderableRectangle *r = &rectangles[i];
